+++ doc/whats_new.rst
@@ -15,6 +15,14 @@ Changes
 
 - Lighting used for interactive surface plots changed; plots may look a bit
   different.
+- plotting.view_connectome default colormap is `bwr`, consistent with plot_connectome.
+- plotting.view_connectome parameter names are consistent with plot_connectome:
+
+ - coords is now node_coord
+ - marker_size is noe node_size
+ - cmap is now edge_cmap
+ - threshold is now edge_threshold
+
 - plotting.view_markers() default marker color is now 'red', consistent with add_markers().
 - plotting.view_markers() parameter names are consistent with add_markers():